file model update

+++ app/django/apiV1/serializers/docs.py
@@ -1,4 +1,5 @@
 import json
+import os
 from urllib.parse import urlparse
 
 from django.conf import settings
@@ -199,14 +200,16 @@ class DocumentSerializer(serializers.ModelSerializer):
         user = request.user
 
         new_links = self.initial_data.getlist('newLinks', [])  # Links 처리
-        for link in new_links:
-            Link.objects.create(docs=docs, link=validate_link(link))
+        if new_links:
+            for link in new_links:
+                Link.objects.create(docs=docs, link=validate_link(link))
 
         # Files 처리
         new_files = request.FILES.getlist('new_files')
         new_descs = request.data.getlist('new_descs')
-        for file, desc in zip(new_files, new_descs):
-            File.objects.create(docs=docs, file=file, description=desc, user=user)
+        if new_files:
+            for file, desc in zip(new_files, new_descs):
+                File.objects.create(docs=docs, file=file, description=desc, user=user)
         return docs
 
     @transaction.atomic
@@ -222,49 +225,53 @@ class DocumentSerializer(serializers.ModelSerializer):
         try:
             # --- Links 처리 ---
             old_links = self.initial_data.getlist('links', [])
-            for json_link in old_links:
-                link = json.loads(json_link)
-                link_object = Link.objects.get(pk=link.get('pk'))
-                if link.get('del'):
-                    link_object.delete()
-                else:
-                    link_object.link = validate_link(link.get('link'))
-                    link_object.save()
+            if old_links:
+                for json_link in old_links:
+                    link = json.loads(json_link)
+                    link_object = Link.objects.get(pk=link.get('pk'))
+                    if link.get('del'):
+                        link_object.delete()
+                    else:
+                        link_object.link = validate_link(link.get('link'))
+                        link_object.save()
 
             new_links = self.initial_data.getlist('newLinks', [])
-            for link in new_links:
-                Link.objects.create(docs=instance, link=validate_link(link))
+            if new_links:
+                for link in new_links:
+                    Link.objects.create(docs=instance, link=validate_link(link))
 
             # --- Files 처리 ---
             new_files = request.FILES.getlist('new_files')
             new_descs = request.data.getlist('new_descs')
 
-            for file, desc in zip(new_files, new_descs):
-                File.objects.create(docs=instance, file=file, description=desc, user=user)
+            if new_files:
+                for file, desc in zip(new_files, new_descs):
+                    File.objects.create(docs=instance, file=file, description=desc, user=user)
 
             old_files = self.initial_data.getlist('files', [])
             cng_pks = self.initial_data.getlist('cngPks', [])
             cng_files = self.initial_data.getlist('cngFiles', [])
             cng_maps = dict(zip(cng_pks, cng_files))
 
-            for json_file in old_files:
-                file = json.loads(json_file)
-                file_object = File.objects.get(pk=file.get('pk'))
+            if old_files:
+                for json_file in old_files:
+                    file = json.loads(json_file)
+                    file_object = File.objects.get(pk=file.get('pk'))
 
-                if file.get('del'):
-                    file_object.delete()
-                    continue
+                    if file.get('del'):
+                        file_object.delete()
+                        continue
 
-                cng_file = cng_maps.get(str(file.get('pk')))
-                if cng_file:
-                    try:
-                        if default_storage.exists(file_object.file.name):
-                            default_storage.delete(file_object.file.name)
-                    except Exception as e:
-                        print(f"파일 처리 중 오류 발생: {e}")
-                    file_object.file = cng_file
-                    file_object.user = user
-                    file_object.save()
+                    cng_file = cng_maps.get(str(file.get('pk')))
+                    if cng_file:
+                        try:
+                            if default_storage.exists(file_object.file.name):
+                                default_storage.delete(file_object.file.name)
+                        except Exception as e:
+                            print(f"파일 처리 중 오류 발생: {e}")
+                        file_object.file = cng_file
+                        file_object.user = user
+                        file_object.save()
 
         except Exception as e:
             print(f"링크 및 파일 처리 중 오류 발생: {e}")
